2_DHT servers should be on the hashtable
