+++ server/DHT/distributed_hash_table.go
@@ -1,0 +1,6 @@
+package DHT
+
+type dht interface {
+	Who()
+	Update()
+}
