+++ server/DHT/next_iteration_dht.go
@@ -1,0 +1,7 @@
+package DHT
+
+//goal  is who ~ O(1)
+// who under read lock
+// update under write lock
+type n_i_dht struct {
+}
